Rename e2e_tests, just to tests, its on lib-level already, fix lib exports
