+++ aquadoggo/src/db/stores/test_utils/runner.rs
@@ -11,12 +11,13 @@ use rstest::fixture;
 use tokio::runtime::Builder;
 use tokio::sync::Mutex;
 
+use crate::config::Configuration;
 use crate::context::Context;
 use crate::db::provider::SqlStorage;
 use crate::db::stores::test_utils::{TestData, TestDatabase};
 use crate::db::Pool;
+use crate::schema::SchemaProvider;
 use crate::test_helpers::{initialize_db, initialize_db_with_url};
-use crate::{Configuration, SchemaProvider};
 
 use super::{doggo_fields, doggo_schema};
